MNMS-414 refactor: api 분리
