+++ src/main/java/com/tekcit/festival/domain/user/dto/response/AssignmentDTO.java
@@ -3,6 +3,8 @@ package com.tekcit.festival.domain.user.dto.response;
 import com.tekcit.festival.domain.user.entity.User;
 import com.tekcit.festival.domain.user.entity.UserProfile;
 import io.swagger.v3.oas.annotations.media.Schema;
+import jakarta.validation.constraints.NotBlank;
+import jakarta.validation.constraints.Pattern;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
@@ -14,17 +16,18 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @Builder
 public class AssignmentDTO {
-    @Schema(description = "양도 시 사용자 이름")
+    @Schema(description = "사용자 이름")
     private String name;
 
-    @Schema(description = "양도 시 사용자 전화번호")
-    private String phone;
+    @Schema(description = "회원 주민번호 앞자리 + 뒷자리 첫자리")
+    private String residentNum;
 
     @Schema(hidden = true)
     public static AssignmentDTO fromUserEntity(User user) {
+        UserProfile userProfile = user.getUserProfile();
         return AssignmentDTO.builder()
                 .name(user.getName())
-                .phone(user.getPhone())
+                .residentNum(userProfile.getResidentNum())
                 .build();
     }
 }
