+++ src/main/java/com/tekcit/festival/domain/user/repository/UserRepository.java
@@ -5,6 +5,8 @@ import com.tekcit.festival.domain.user.enums.OAuthProvider;
 import com.tekcit.festival.domain.user.enums.UserRole;
 import org.springframework.data.jpa.repository.EntityGraph;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Query;
+import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 
 import java.util.List;
@@ -31,4 +33,7 @@ public interface UserRepository extends JpaRepository<User, Long> {
 
     @EntityGraph(attributePaths = {"hostProfile"})
     List<User> findAllHostByRole(UserRole userRole);
+
+    @Query("select u from User u join fetch u.userProfile where u.email = :email")
+    Optional<User> findByEmailWithProfile(@Param("email") String email);
 }
